requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
psycopg2-binary==2.9.9
faiss-cpu==1.7.4
numpy==1.26.3
python-dotenv==1.0.0
